legacy implicit to provide binary compatible jenum enum
