+++ core/jvm/src/main/scala/scalaz/std/java/Enum.scala
@@ -4,10 +4,14 @@ package std.java
 import scala.reflect.ClassTag
 
 trait EnumInstances {
+  // legacy low(er) priority implicit for binary compatibility
+  implicit def enumInstance[E <: java.lang.Enum[E]]: Equal[E] = Equal.equal[E](_ eq _)
+
   implicit def enumInstance[E <: java.lang.Enum[E] : ClassTag]: Enum[E] = new Enum[E] {
     private[this] lazy val values: Array[E] = implicitly[ClassTag[E]].runtimeClass.asInstanceOf[Class[E]].getEnumConstants
 
     override def equal(x: E, y: E): Boolean = x eq y
+    override def equalIsNatural: Boolean = true
 
     override def pred(e: E): E  = values((values.indexOf(e) + values.length - 1) % values.length)
     override def succ(e: E): E  = values((values.indexOf(e) + 1) % values.length)
