+++ core/native/src/main/scala/scalaz/std/java/Enum.scala
@@ -1,0 +1,8 @@
+package scalaz
+package std.java
+
+trait EnumInstances {
+  implicit def enumInstance[E <: java.lang.Enum[E]]: Equal[E] = Equal.equal[E](_ eq _)
+}
+
+object enum extends EnumInstances
